- Add/Remove/Rename most sans system tests files as well as corresponding CMakeList text files. - Most files with suffix _V2 (meaning new command interface) have been strip of the suffix as the original tests (test with same name but no V2) use the deprecated interface, and are mostly redundant with respect to V2 - Fix tests that can use the new command interface and are broken
