+++ Testing/SystemTests/tests/framework/ISIS/SANS/LARMOR/SANSLARMORMultiPeriodAddFilesTest.py
@@ -5,45 +5,50 @@
 #   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 # SPDX - License - Identifier: GPL - 3.0 +
 # pylint: disable=no-init
-import os
+
 import systemtesting
+import os
+
 from ISIS.SANS.isis_sans_system_test import ISISSansSystemTest
 from mantid.kernel import config
-from mantid.simpleapi import DeleteWorkspace
-
-from ISISCommandInterface import AssignSample, DefaultTrans, Detector, Gravity, LARMOR, MaskFile, Set1D, WavRangeReduction
+from mantid.api import AnalysisDataService
+from sans.command_interface.ISISCommandInterface import (
+    Set1D,
+    Detector,
+    MaskFile,
+    Gravity,
+    AssignSample,
+    WavRangeReduction,
+    DefaultTrans,
+    UseCompatibilityMode,
+    AddRuns,
+    LARMOR,
+)
 from sans.common.enums import SANSInstrument
-from SANSadd2 import add_runs
 
 
 @ISISSansSystemTest(SANSInstrument.LARMOR)
-class LARMORMultiPeriodAddEventFiles(systemtesting.MantidSystemTest):
+class LARMORMultiPeriodAddEventFilesTest(systemtesting.MantidSystemTest):
     def requiredMemoryMB(self):
         """Requires 2.5Gb"""
         return 2500
 
     def runTest(self):
+        UseCompatibilityMode()
         LARMOR()
         Set1D()
         Detector("DetectorBench")
         MaskFile("USER_LARMOR_151B_LarmorTeam_80tubes_BenchRot1p4_M4_r3699.txt")
         Gravity(True)
-        add_runs(("13065", "13065"), "LARMOR", "nxs", lowMem=True)
+        AddRuns(("13065", "13065"), "LARMOR", "nxs", lowMem=True)
 
         AssignSample("13065-add.nxs")
         WavRangeReduction(2, 4, DefaultTrans)
 
         # Clean up
-        to_clean = [
-            "13065_sans_nxs",
-            "13065p1rear_1D_2.0_4.0_incident_monitor",
-            "13065p2rear_1D_2.0_4.0_incident_monitor",
-            "13065p3rear_1D_2.0_4.0_incident_monitor",
-            "13065p4rear_1D_2.0_4.0_incident_monitor",
-            "80tubeCalibration_1-05-2015_r3157-3160",
-        ]
-        for workspace in to_clean:
-            DeleteWorkspace(workspace)
+        for element in AnalysisDataService.getObjectNames():
+            if AnalysisDataService.doesExist(element) and element != "13065_p1rear_1D_2.0_4.0":
+                AnalysisDataService.remove(element)
 
         paths = [
             os.path.join(config["defaultsave.directory"], "LARMOR00013065-add.nxs"),
@@ -61,4 +66,4 @@ class LARMORMultiPeriodAddEventFiles(systemtesting.MantidSystemTest):
         self.disableChecking.append("Instrument")
         self.disableChecking.append("Axes")
 
-        return "13065p1rear_1D_2.0_4.0", "LARMORMultiPeriodAddEventFiles.nxs"
+        return "13065_p1rear_1D_2.0_4.0", "LARMORMultiPeriodAddEventFiles.nxs"
