+++ Testing/SystemTests/tests/framework/ISIS/SANS/LOQ/CMakeLists.txt
@@ -1,17 +1,13 @@
 # System tests related to the LOQ instrument at ISIS.
 
 set(SYSTEST_NAMES
-    LOQReloadWorkspaces.py
     SANSLOQAddBatch.py
     SANSLOQBatch.py
-    SANSLOQBatch_V2.py
     SANSLOQCan2D.py
     SANSLOQCan2D_TOML.py
-    SANSLOQCan2D_V2.py
     SANSLOQCentreNoGrav.py
     SANSLOQLegacyUserFile.py
-    SANSLOQReductionGUI.py
-    SANSLOQReductionGUITest_V2.py
+    SANSLOQReductionGUITest.py
     SANSLOQTransFitWorkspace2D.py
 )
