+++ Testing/SystemTests/tests/framework/ISIS/SANS/LOQ/SANSLOQBatch.py
@@ -7,28 +7,25 @@
 # pylint: disable=no-init
 
 import systemtesting
-from ISIS.SANS.isis_sans_system_test import ISISSansSystemTest
-from mantid.simpleapi import LoadNexus, Plus
-from mantid import config, FileFinder
-from ISISCommandInterface import Detector, Gravity, LOQ, MaskFile, Set1D
-from SANSBatchMode import BatchReduce
 import os.path
 
+from ISIS.SANS.isis_sans_system_test import ISISSansSystemTest
+from mantid.kernel import config
+from mantid.api import FileFinder
+from mantid.simpleapi import LoadNexus, Plus
+from sans.command_interface.ISISCommandInterface import LOQ, Detector, Set1D, MaskFile, Gravity, BatchReduce
 from sans.common.enums import SANSInstrument
 
 
 @ISISSansSystemTest(SANSInstrument.LOQ)
-class SANSLOQBatch(systemtesting.MantidSystemTest):
+class SANSLOQBatchTest(systemtesting.MantidSystemTest):
     def runTest(self):
         # DataPath("../Data/LOQ/")
         # UserPath("../Data/LOQ/")
 
         # here we are testing the LOQ setup
         LOQ()
-        # rear detector
         Detector("main-detector-bank")
-        # test batch mode, although only the analysis from the last line is checked
-        # Find the file , this should really be in the BatchReduce reduction step
         csv_file = FileFinder.getFullPath("batch_input.csv")
 
         Set1D()
@@ -37,18 +34,15 @@ class SANSLOQBatch(systemtesting.MantidSystemTest):
 
         BatchReduce(csv_file, "raw", plotresults=False, saveAlgs={"SaveCanSAS1D": "xml", "SaveNexus": "nxs"})
 
-        LoadNexus(Filename="54433sans.nxs", OutputWorkspace="result")
-        Plus(LHSWorkspace="result", RHSWorkspace="99630sanotrans", OutputWorkspace="result")
+        LoadNexus(Filename="54433sans_main_1D_2.2_10.0.nxs", OutputWorkspace="result")
+        Plus(LHSWorkspace="result", RHSWorkspace="99630sanotrans_main_1D_2.2_10.0", OutputWorkspace="result")
 
-        os.remove(os.path.join(config["defaultsave.directory"], "54433sans.nxs"))
-        os.remove(os.path.join(config["defaultsave.directory"], "99630sanotrans.nxs"))
-        os.remove(os.path.join(config["defaultsave.directory"], "54433sans.xml"))
-        os.remove(os.path.join(config["defaultsave.directory"], "99630sanotrans.xml"))
+        os.remove(os.path.join(config["defaultsave.directory"], "54433sans_main_1D_2.2_10.0.nxs"))
+        os.remove(os.path.join(config["defaultsave.directory"], "99630sanotrans_main_1D_2.2_10.0.nxs"))
+        os.remove(os.path.join(config["defaultsave.directory"], "54433sans_main_1D_2.2_10.0.xml"))
+        os.remove(os.path.join(config["defaultsave.directory"], "99630sanotrans_main_1D_2.2_10.0.xml"))
 
     def validate(self):
-        # Need to disable checking of the Spectra-Detector map because it isn't
-        # fully saved out to the nexus file (it's limited to the spectra that
-        # are actually present in the saved workspace).
         self.disableChecking.append("SpectraMap")
         self.disableChecking.append("Axes")
         self.disableChecking.append("Instrument")
