+++ Testing/SystemTests/tests/framework/ISIS/SANS/LOQ/SANSLOQTransFitWorkspace2D.py
@@ -9,7 +9,7 @@
 import systemtesting
 from ISIS.SANS.isis_sans_system_test import ISISSansSystemTest
 from mantid.simpleapi import DeleteWorkspace, RenameWorkspace
-from ISISCommandInterface import (
+from sans.command_interface.ISISCommandInterface import (
     AssignCan,
     AssignSample,
     Detector,
