+++ Testing/SystemTests/tests/framework/ISIS/SANS/SANS2D/SANS2DFrontNoGrav.py
@@ -5,16 +5,26 @@
 #   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 # SPDX - License - Identifier: GPL - 3.0 +
 # pylint: disable=no-init
-
 import systemtesting
+import mantid  # noqa
 from ISIS.SANS.isis_sans_system_test import ISISSansSystemTest
-from ISISCommandInterface import AssignSample, Gravity, MaskFile, SANS2D, Set1D, SetDetectorOffsets, WavRangeReduction
+from sans.command_interface.ISISCommandInterface import (
+    SANS2D,
+    MaskFile,
+    SetDetectorOffsets,
+    Gravity,
+    Set1D,
+    UseCompatibilityMode,
+    AssignSample,
+    WavRangeReduction,
+)
 from sans.common.enums import SANSInstrument
 
 
 @ISISSansSystemTest(SANSInstrument.SANS2D)
-class SANS2DFrontNoGrav(systemtesting.MantidSystemTest):
+class SANS2DFrontNoGravTest(systemtesting.MantidSystemTest):
     def runTest(self):
+        UseCompatibilityMode()
         SANS2D()
         MaskFile("MASKSANS2D_094i_RKH.txt")
         SetDetectorOffsets("REAR", -16.0, 58.0, 0.0, 0.0, 0.0, 0.0)
@@ -23,18 +33,20 @@ class SANS2DFrontNoGrav(systemtesting.MantidSystemTest):
         Set1D()
 
         AssignSample("2500.nxs")
+
         WavRangeReduction(4.6, 12.85, False)
 
     def validate(self):
         self.disableChecking.append("SpectraMap")
         self.disableChecking.append("Axes")
         self.disableChecking.append("Instrument")
-        return "2500front_1D_4.6_12.85", "SANS2DFrontNoGrav.nxs"
+        return "2500_front_1D_4.6_12.85", "SANS2DFrontNoGrav.nxs"
 
 
 @ISISSansSystemTest(SANSInstrument.SANS2D)
-class SANS2DWithExtraLengthGravity(systemtesting.MantidSystemTest):
+class SANS2DWithExtraLengthGravityTest(systemtesting.MantidSystemTest):
     def runTest(self):
+        UseCompatibilityMode()
         SANS2D()
         MaskFile("MASKSANS2D_094i_RKH.txt")
         SetDetectorOffsets("REAR", -16.0, 58.0, 0.0, 0.0, 0.0, 0.0)
@@ -50,4 +62,4 @@ class SANS2DWithExtraLengthGravity(systemtesting.MantidSystemTest):
         self.disableChecking.append("SpectraMap")
         self.disableChecking.append("Axes")
         self.disableChecking.append("Instrument")
-        return "2500front_1D_4.6_12.85", "SANS2DWithExtraLengthGravity.nxs"
+        return "2500_front_1D_4.6_12.85", "SANS2DWithExtraLengthGravity.nxs"
