+++ Testing/SystemTests/tests/framework/ISIS/SANS/SANS2D/SANS2DLimitEventsTime.py
@@ -7,14 +7,16 @@
 # pylint: disable=no-init
 
 import systemtesting
+import mantid  # noqa
 from ISIS.SANS.isis_sans_system_test import ISISSansSystemTest
-from ISISCommandInterface import AssignSample, MaskFile, SANS2D, WavRangeReduction
+from sans.command_interface.ISISCommandInterface import SANS2D, MaskFile, AssignSample, WavRangeReduction, UseCompatibilityMode
 from sans.common.enums import SANSInstrument
 
 
 @ISISSansSystemTest(SANSInstrument.SANS2D)
-class SANS2DLimitEventsTime(systemtesting.MantidSystemTest):
+class SANS2DLimitEventsTimeTest(systemtesting.MantidSystemTest):
     def runTest(self):
+        UseCompatibilityMode()
         SANS2D()
         MaskFile("MaskSANS2DReductionGUI_LimitEventsTime.txt")
         AssignSample("22048")
@@ -24,4 +26,4 @@ class SANS2DLimitEventsTime(systemtesting.MantidSystemTest):
         self.disableChecking.append("SpectraMap")
         self.disableChecking.append("Axes")
         self.disableChecking.append("Instrument")
-        return "22048rear_1D_1.5_12.5", "SANSReductionGUI_LimitEventsTime.nxs"
+        return "22048_rear_1D_1.5_12.5", "SANSReductionGUI_LimitEventsTime.nxs"
