+++ Testing/SystemTests/tests/framework/ISIS/SANS/WORKFLOWS/SANSQResolutionTest.py
@@ -8,49 +8,32 @@
 
 import systemtesting
 from ISIS.SANS.isis_sans_system_test import ISISSansSystemTest
-from ISISCommandInterface import (
-    set_q_resolution_a1,
-    set_q_resolution_a2,
-    set_q_resolution_collimation_length,
-    set_q_resolution_delta_r,
-    set_q_resolution_h1,
-    set_q_resolution_h2,
-    set_q_resolution_moderator,
-    set_q_resolution_use,
-    set_q_resolution_w1,
-    set_q_resolution_w2,
-    AssignSample,
-    Gravity,
-    MaskFile,
-    SANS2D,
-    Set1D,
-    SetDetectorOffsets,
-    WavRangeReduction,
-)
+from sans.command_interface.ISISCommandInterface import AssignSample, Gravity, MaskFile, SANS2D, Set1D, SetDetectorOffsets, director
 from sans.common.enums import SANSInstrument
+
+tolerance = 1e-6
 
 
 @ISISSansSystemTest(SANSInstrument.SANS2D)
 class SANSQResolutionWithoutGravity(systemtesting.MantidSystemTest):
     def runTest(self):
+
         SANS2D()
-        MaskFile("MASKSANS2D_094i_RKH.txt")
+        # MaskFile contains setting for Q Resolution
+        MaskFile("/home/twy67886/Documents/MASKSANS2D_094i_RKH_QRes1.txt")
         SetDetectorOffsets("REAR", -16.0, 58.0, 0.0, 0.0, 0.0, 0.0)
         SetDetectorOffsets("FRONT", -44.0, -20.0, 47.0, 0.0, 1.0, 1.0)
         Gravity(False)
         Set1D()
         AssignSample("2500.nxs")
-
-        # Provide settings for QResolution
-        set_q_resolution_use(True)
-        moderator_file_name = "ModeratorStdDev_TS2_SANS_30Jul2015.txt"
-        set_q_resolution_a1(a1=2)
-        set_q_resolution_a2(a2=3)
-        set_q_resolution_delta_r(delta_r=2)
-        set_q_resolution_collimation_length(collimation_length=10)
-        set_q_resolution_moderator(file_name=moderator_file_name)
-
-        WavRangeReduction(4.6, 12.85, False)
+        state = director.process_commands()
+        q = state.convert_to_q
+        self.assertTrue(q.use_q_resolution)
+        self.assertDelta(q.q_resolution_a1, 0.002, tolerance)
+        self.assertDelta(q.q_resolution_a2, 0.003, tolerance)
+        self.assertDelta(q.q_resolution_delta_r, 0.002, tolerance)
+        self.assertDelta(q.q_resolution_collimation_length, 10.0, tolerance)
+        self.assertEqual(q.moderator_file, "ModeratorStdDev_TS2_SANS_30Jul2015.txt")
 
     def validate(self):
         self.disableChecking.append("Instrument")
@@ -61,26 +44,25 @@ class SANSQResolutionWithoutGravity(systemtesting.MantidSystemTest):
 class SANSQResolutionWithGravity(systemtesting.MantidSystemTest):
     def runTest(self):
         SANS2D()
-        MaskFile("MASKSANS2D_094i_RKH.txt")
+        # MaskFile contains setting for Q Resolution
+        MaskFile("/home/twy67886/Documents/MASKSANS2D_094i_RKH_QRes2.txt")
         SetDetectorOffsets("REAR", -16.0, 58.0, 0.0, 0.0, 0.0, 0.0)
         SetDetectorOffsets("FRONT", -44.0, -20.0, 47.0, 0.0, 1.0, 1.0)
         Gravity(flag=True, extra_length=10.0)
         Set1D()
         AssignSample("2500.nxs")
-
-        # Provide settings for QResolution
-        set_q_resolution_use(True)
-        moderator_file_name = "ModeratorStdDev_TS2_SANS_30Jul2015.txt"
-        set_q_resolution_h1(h1=2)
-        set_q_resolution_w1(w1=3)
-        set_q_resolution_h2(h2=4)
-        set_q_resolution_w2(w2=5)
-
-        set_q_resolution_delta_r(delta_r=2)
-        set_q_resolution_collimation_length(collimation_length=5)
-        set_q_resolution_moderator(file_name=moderator_file_name)
-
-        WavRangeReduction(4.6, 12.85, False)
+        state = director.process_commands()
+        q = state.convert_to_q
+        self.assertTrue(q.use_q_resolution)
+        self.assertDelta(q.q_resolution_h1, 0.002, tolerance)
+        self.assertDelta(q.q_resolution_h2, 0.004, tolerance)
+        self.assertDelta(q.q_resolution_w1, 0.003, tolerance)
+        self.assertDelta(q.q_resolution_w2, 0.005, tolerance)
+        self.assertDelta(q.q_resolution_delta_r, 0.002, tolerance)
+        self.assertDelta(q.q_resolution_collimation_length, 5.0, tolerance)
+        self.assertEqual(q.moderator_file, "ModeratorStdDev_TS2_SANS_30Jul2015.txt")
+        self.assertTrue(q.use_gravity)
+        self.assertDelta(q.gravity_extra_length, 10.0, tolerance)
 
     def validate(self):
         self.disableChecking.append("Instrument")
