+++ Testing/SystemTests/tests/framework/ISIS/SANS/WORKFLOWS/SANSSaveTest.py
@@ -326,7 +326,7 @@ class SANSSaveRunnerTest(systemtesting.MantidSystemTest):
 
     def runTest(self):
         suite = unittest.TestSuite()
-        suite.addTest(unittest.makeSuite(SANSSaveTest, "test"))
+        suite.addTest(unittest.defaultTestLoader.loadTestsFromTestCase(SANSSaveTest))
         runner = unittest.TextTestRunner()
         res = runner.run(suite)
         if res.wasSuccessful():
